move to lib and minor refactor
